Use a double for scaling instead of an integer
All bitmap drivers now support arbitrary scaling, and the scaling
factor is a double, allowing arbitrary window scaling in all bitmap
modes (making nelgin happy).

While we're here, fix bugs in horizontal interpolation and X window
resizeing.

+++ src/conio/ansi_cio.c
@@ -970,7 +970,7 @@ int ansi_initio_cb(void)
 #if defined(__BORLANDC__)
         #pragma argsused
 #endif
-int ansi_initciolib(long inmode)
+int ansi_initciolib(int inmode)
 {
 	int i;
 	char *init="\033[s\033[99B\033[99B\033[99B_\033[99C\033[99C\033[99C_\033[6n\033[u\033[0m_\033[2J\033[H";
