+++ src/conio/ansi_cio.h
@@ -63,7 +63,7 @@ int ansi_wherey(void);
 int ansi_wherex(void);
 int ansi_putch(int ch);
 void ansi_gotoxy(int x, int y);
-int ansi_initciolib(long inmode);
+int ansi_initciolib(int inmode);
 void ansi_gettextinfo(struct text_info *info);
 void ansi_setcursortype(int type);
 int ansi_getch(void);
