+++ src/conio/win32gdi.h
@@ -19,7 +19,7 @@ void gdi_flush(void);
 int gdi_mousepointer(enum ciolib_mouse_ptr type);
 void gdi_setwinposition(int x, int y);
 void gdi_setwinsize(int w, int h);
-int gdi_getscaling(void);
-void gdi_setscaling(int newval);
+double gdi_getscaling(void);
+void gdi_setscaling(double newval);
 
 #endif
